feat: migrate from deprecated create-react-app to Vite
