+++ packages/backend/sample.env
@@ -15,3 +15,4 @@ SENDGRID_GROUP_ID=21140
 
 S3_ID=AWS S3 ID for photo upload
 S3_SECRET=AWS S3 Secret for photo upload
+PORT=5000
