+++ packages/backend/tsconfig.json
@@ -11,9 +11,7 @@
     "strict": true,
     "types": [
       "jest"
-    ],
-    "baseUrl": ".",
-    "paths": { }
+    ]
   },
   "include": ["./src/**/*"]
 }
