+++ packages/frontend/package.json
@@ -2,20 +2,13 @@
   "name": "star-vote",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "dependencies": {
     "@date-io/date-fns": "^1.3.13",
-    "@emotion/react": "^11.10.4",
-    "@emotion/styled": "^11.10.4",
-    "@mui/icons-material": "^5.10.9",
-    "@mui/material": "^5.10.10",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/jest": "^27.4.1",
-    "@types/luxon": "^3.3.0",
-    "@types/node": "^17.0.23",
-    "@types/react": "^17.0.44",
-    "@types/react-dom": "^17.0.15",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.15.12",
+    "@mui/material": "^5.15.12",
     "@ungap/structured-clone": "0.3.4",
     "date-fns": "^2.28.0",
     "jwt-decode": "^3.1.2",
@@ -28,26 +21,22 @@
     "react-ratings-declarative": "3.4.1",
     "react-router": "6.0.2",
     "react-router-dom": "6.0.2",
-    "react-scripts": "^5.0.1",
     "recharts": "^2.6.2",
     "rewire": "^7.0.0",
-    "typescript": "^3.9.10",
     "uuid": "^9.0.1",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
-    "start": "node ./start.js",
-    "build": "node ./build.js",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev": "vite",
+    "dev:clean": "npm run dev -- --force",
+    "build": "tsc && vite build",
+    "start": "vite preview"
   },
   "eslintConfig": {
     "extends": [
-      "react-app",
-      "react-app/jest"
+      "react-app"
     ]
   },
-  "proxy": "http://localhost:5000/",
   "browserslist": {
     "production": [
       ">0.2%",
@@ -59,5 +48,17 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@types/luxon": "^3.3.0",
+    "@types/node": "^20.11.24",
+    "@types/react": "^17.0.44",
+    "@types/react-dom": "^17.0.15",
+    "@vitejs/plugin-react": "^4.2.1",
+    "identity-obj-proxy": "^3.0.0",
+    "typescript": "^5.3.3",
+    "vite": "^5.1.5",
+    "vite-plugin-checker": "^0.6.4",
+    "vite-plugin-env-compatible": "^2.0.1"
   }
 }
