+++ packages/frontend/src/components/Election/Voting/StarBallotView.jsx
@@ -1,50 +1,53 @@
 import React, { useContext } from "react";
-import GenericBallotView from "./GenericBallotView.js";
+import GenericBallotView from "./GenericBallotView";
 import Typography from '@mui/material/Typography';
 import { BallotContext } from "./VotePage";
 
-function scoresAreUnderVote({ scores }) {
+function scoresAreUnderVote({scores}){
   let five_selected = false
   let zero_selected = false
   let all_null = true
-  for (let i = 0; i < scores.length; i++) {
-    if (scores[i] != null) all_null = false
-    if (scores[i] == null || scores[i] == 0) zero_selected = true
-    if (scores[i] == 5) five_selected = true
+  for(let i = 0; i < scores.length; i++){
+    if(scores[i] != null) all_null = false
+    if(scores[i] == null || scores[i] == 0) zero_selected = true
+    if(scores[i] == 5) five_selected = true
   }
   return !(all_null || (five_selected && zero_selected))
 }
 
 // Renders a complete RCV ballot for a single race
-export default function StarPRBallotView({
-  race,
-  candidates,
-  onUpdate
-}) {
+export default function StarBallotView() {
   const ballotContext = useContext(BallotContext);
+
   const instructions = (
     <>
       <Typography align='left' sx={{ typography: { sm: 'body1', xs: 'body2' } }} component="li">
         Give your favorite(s) five stars.
       </Typography>
       <Typography align='left' sx={{ typography: { sm: 'body1', xs: 'body2' } }} component="li">
-        Give your last choice(s) zero stars.
+        Give your last choice(s) zero stars or leave blank.
       </Typography>
       <Typography align='left' sx={{ typography: { sm: 'body1', xs: 'body2' } }} component="li">
-        Show preference order and level of support.
+        Score other candidates as desired.
       </Typography>
       <Typography align='left' sx={{ typography: { sm: 'body1', xs: 'body2' } }} component="li">
         Equal scores indicate no preference.
       </Typography>
-      <Typography align='left' sx={{ typography: { sm: 'body1', xs: 'body2' } }} component="li">
-        Those left blank receive zero stars.
-      </Typography>
     </>
   )
   const footer = (
-    <Typography align='center' component="p">
-      Winners in Proportional STAR Voting are selected in rounds. Each round elects the candidate with the highest total score, then designates that candidate's strongest supporters are represented. Subsequent rounds include all voters who are not yet fully represented.
-    </Typography>
+    <>
+    {ballotContext.race.num_winners == 1 &&
+      <Typography align='center' component="p">
+        The two highest scoring candidates are finalists.<br/>Your full vote goes to the finalist you prefer.
+      </Typography>
+    }
+    {ballotContext.race.num_winners > 1 && 
+      <Typography align='center' component="p">
+        {`This election uses STAR Voting and will elect ${ballotContext.race.num_winners} winners. In STAR Voting the two highest scoring candidates are finalists and the finalist preferred by more voters wins.`}
+      </Typography>
+    }
+    </>
   )
   let warning = null;
 
@@ -61,9 +64,9 @@ export default function StarPRBallotView({
     <GenericBallotView
       key="starBallot"
       columns={[0, 1, 2, 3, 4, 5]}
-      methodName="STAR PR Voting"
+      methodName="STAR Voting"
       instructions={instructions}
-      learnMoreLink='https://www.starvoting.org/star-pr'
+      learnMoreLink="https://www.starvoting.org/star"
       leftTitle='Worst'
       rightTitle='Best'
       onClick={(i, j) => {
